+++ packages/spacecat-shared-http-utils/src/index.js
@@ -15,6 +15,7 @@ import { Response } from '@adobe/fetch';
 import LegacyApiKeyHandler from './auth/handlers/legacy-api-key.js';
 import AdobeImsHandler from './auth/handlers/ims.js';
 import ScopedApiKeyHandler from './auth/handlers/scoped-api-key.js';
+import JwtHandler from './auth/handlers/jwt.js';
 
 const HEADER_CONTENT_TYPE = 'content-type';
 const HEADER_ERROR = 'x-error';
@@ -106,4 +107,6 @@ export { authWrapper } from './auth/auth-wrapper.js';
 export { enrichPathInfo } from './enrich-path-info-wrapper.js';
 export { hashWithSHA256 } from './auth/generate-hash.js';
 
-export { AdobeImsHandler, ScopedApiKeyHandler, LegacyApiKeyHandler };
+export {
+  AdobeImsHandler, ScopedApiKeyHandler, LegacyApiKeyHandler, JwtHandler,
+};